services/trade-service/requirements.txt
 flask
  gunicorn
  psycopg2-binary
  boto3
  python-dotenv
  pytest
  pytest-mock
